refactor : Url 엔티티 관련 수정
- Algorithm enum 삭제
- Url 엔티티 제약 조건, setShortUrl 추가
