--- src/main/java/com/programmers/springbooturlshortener/domain/Algorithm.java
@@ -1,7 +1,0 @@
-package com.programmers.springbooturlshortener.domain;
-
-public enum Algorithm {
-
-	BASE62,
-	BASE64
-}
