+++ src/main/java/com/programmers/springbooturlshortener/domain/url/Url.java
@@ -1,10 +1,14 @@
-package com.programmers.springbooturlshortener.domain;
-
-import javax.persistence.*;
+package com.programmers.springbooturlshortener.domain.url;
 
 import lombok.AccessLevel;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.lang.Nullable;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
 @Entity
 @Table(
@@ -17,19 +21,22 @@ import lombok.NoArgsConstructor;
         }
 )
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
 public class Url {
 
     @Id
     @GeneratedValue
     private Long id;
 
-    @Enumerated(EnumType.STRING)
+    @NotNull
     @Column(name = "algorithm", length = 20)
-    private Algorithm algorithm;
+    private String algorithm;
 
-    @Column(name = "origin_url", length = 2000)
+    @NotBlank
+    @Column(name = "origin_url", nullable = false, length = 2000)
     private String originUrl;
 
+    @Nullable
     @Column(name = "short_url", length = 7)
     private String shortUrl;
 
@@ -37,14 +44,17 @@ public class Url {
     private Long requestCount;
 
     @Builder
-    public Url(Algorithm algorithm, String originUrl, String shortUrl, Long requestCount) {
+    public Url(String algorithm, String originUrl, Long requestCount) {
         this.algorithm = algorithm;
         this.originUrl = originUrl;
-        this.shortUrl = shortUrl;
         this.requestCount = requestCount;
     }
 
     public void increaseRequestCount() {
         this.requestCount++;
+    }
+
+    public void setShortUrl(String shortUrl) {
+        this.shortUrl = shortUrl;
     }
 }
